move attributionInfo to map/components
