+++ src/index.html
@@ -11,6 +11,7 @@
   <ba-header></ba-header>
   <ba-ol-map></ba-ol-map>
   <ba-map-button-container></ba-map-button-container>
+  <ba-attribution-info></ba-attribution-info>
   <ba-footer></ba-footer>
   <!-- <ba-side-panel></ba-side-panel> -->
   <ba-content-panel></ba-content-panel>
