+++ src/modules/footer/components/mapInfo/mapInfo.css
@@ -17,10 +17,3 @@
     max-width: 30em;
     max-height: 3em;
 } 
-
-.attribution-info {
-    display: block;
-    margin-left: 1.5em;
-    max-width: 30em;
-    max-height: 3em;
-}
