+++ src/modules/footer/index.js
@@ -2,7 +2,6 @@ import './i18n';
 import './components/coordinateSelect';
 import './components/mapInfo';
 import './components/baseLayerInfo';
-import './components/attributionInfo';
 import { Footer } from './components/Footer';
 if (!window.customElements.get(Footer.tag)) {
 	window.customElements.define(Footer.tag, Footer);
