+++ src/modules/map/components/attributionInfo/index.js
