+++ src/modules/map/i18n/attributionInfo.provider.js
