+++ src/modules/map/i18n/index.js
@@ -5,6 +5,7 @@ import { provide as geolocationButtonProvide } from './geolocationButton.provide
 import { provide as layerManagerProvide } from './layerManager.provider';
 import { provide as contextMenuProvider } from './contextMenu.provider';
 import { provide as storeProvider } from './store.provider';
+import { provide as attributionInfoProvider } from './attributionInfo.provider'; 
 import { $injector } from '../../../injection';
 const { TranslationService: translationService } = $injector.inject('TranslationService');
 translationService.register('mapProvider', olMapProvide);
@@ -14,3 +15,5 @@ translationService.register('geolocationButtonProvider', geolocationButtonProvid
 translationService.register('layerManagerProvider', layerManagerProvide);
 translationService.register('contextMenuProvider', contextMenuProvider);
 translationService.register('storeProvider', storeProvider);
+translationService.register('attributionInfoProvider', attributionInfoProvider);
+
