+++ src/modules/map/index.js
@@ -6,3 +6,4 @@ import './components/mapButtonsContainer';
 import './components/geolocationButton';
 import './components/contextMenu';
 import './components/olMap';
+import './components/attributionInfo';
