+++ test/modules/footer/components/mapInfo/MapInfo.test.js
@@ -29,7 +29,6 @@ describe('MapInfo', () => {
 
 			expect(element.shadowRoot.querySelector('ba-coordinate-select')).toBeTruthy();
 			expect(element.shadowRoot.querySelector('ba-base-layer-info')).toBeTruthy();
-			expect(element.shadowRoot.querySelector('ba-attribution-info')).toBeTruthy();
 		});
 	});
 });
